chore: [#248] Move CowCard into its own component file
