+++ src/components/CowCard/CowCard.sass
@@ -1,0 +1,1 @@
+.CowCard
