+++ src/components/CowCard/index.js
@@ -1,0 +1,1 @@
+export { CowCard, CowCardSubheader } from './CowCard'
